# docs/00_current/api/mobile-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "AIJobResultEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AIJobResultData"
          },
          "error": {
            "title": "Error",
            "type": "null"
          },
          "ok": {
            "const": true,
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "AIJobResultEnvelope",
        "type": "object"
      },
...
      "AccountDeletionEnvelope": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AccountDeletionData"
          },
          "error": {
            "title": "Error",
            "type": "null"
          },
          "ok": {
            "const": true,
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "AccountDeletionEnvelope",
        "type": "object"
      },
      "AccountDeletionInput": {
        "properties": {
          "confirmation": {
            "title": "Confirmation",
            "type": "string"
          },
          "password": {
            "default": "",
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "confirmation"
        ],
        "title": "AccountDeletionInput",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "MobileBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned screen-oriented API for the My Scoope consumer mobile client.",
    "title": "My Scoope Consumer API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/account/delete": {
      "post": {
        "operationId": "mobile_api_api_delete_account",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeletionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "MobileBearer": []
          }
        ],
        "summary": "Delete Account"
      }
    },
    "/api/v1/ai/jobs/{job_id}": {
      "get": {
        "operationId": "mobile_api_api_ai_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIJobResultEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIJobResultEnvelope"
                }
              }
            },
            "description": "Accepted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "MobileBearer": []
          }
        ],
        "summary": "Ai Job"
      }
    },
    "/api/v1/ai/turns": {
      "post": {
        "operationId": "mobile_api_api_submit_ai_turn",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AITurnInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIJobAcceptedEnvelope"
                }
              }
            },
            "description": "Accepted"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "MobileBearer": []
          }
        ],
        "summary": "Submit Ai Turn"
      }
    },
    "/api/v1/days/{day_id}/meals/{meal_snapshot_key}/check-ins": {
      "post": {
        "operationId": "mobile_api_api_meal_check_in",
        "parameters": [
          {
            "in": "path",
            "name": "day_id",
            "required": true,
            "schema": {
              "title": "Day Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "meal_snapshot_key",
            "required": true,
            "schema": {
              "title": "Meal Snapshot Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealCheckInInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "MobileBearer": []
          }
        ],
        "summary": "Meal Check In"
      }
    },
    "/api/v1/entitlements": {
      "get": {
        "operationId": "mobile_api_api_entitlements",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementsEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "MobileBearer": []
          }
        ],
        "summary": "Entitlements"
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "mobile_api_api_me",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "MobileBearer": []
          }
        ],
        "summary": "Me"
      }
    },
...
  },
  "servers": []
}